requirements.txt
greenlet==3.2.0
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
psycopg2-binary==2.9.10
pytesseract==0.3.13
python-dotenv==1.1.0
SQLAlchemy==2.0.40
typing_extensions==4.13.2
